syscall

Module call

Function definitions

Functions

call_ro

SYS_CALL interface, read-only variant

call_rw

SYS_CALL interface, read-write variant

call_wo

SYS_CALL interface, write-only variant

clock_gettime

Get the current system time

close

Close a file

dup

Copy and transform a file descriptor

dup2

Copy and transform a file descriptor

fchmod

Change file permissions

fchown

Change file ownership

fcntl

Change file descriptor flags

flink

Create a link to a file

fmap⚠

Map a file into memory, but with the ability to set the address to map into, either as a hint or as a requirement of the map.

fpath

Retrieve the canonical path of a file

frename

Rename a file

fstat

Get metadata about a file

fstatvfs

Get metadata about a filesystem

fsync

Sync a file descriptor to its underlying medium

ftruncate

Truncate or extend a file to a specified length

funmap⚠

Unmap whole (or partial) continous memory-mapped files

futex⚠

Fast userspace mutex

futimens

lseek

Seek to offset bytes in a file descriptor

mkns

Make a new scheme namespace

mprotect⚠

Change mapping flags

nanosleep

Sleep for the time specified in req

open

Open a file

openat

Open a file at a specific path

read

Read from a file descriptor into a buffer

rmdir

Remove a directory

sched_yield

Yield the process’s time slice to the kernel

sendfd

Send a file descriptor fd, handled by the scheme providing receiver_socket. flags is currently unused (must be zero), and arg is included in the scheme call.

unlink

Remove a file

write

Write a buffer to a file descriptor