syscall

Module flag

Flags used as an argument to many system calls

Structs

CallFlags

EventFlags

FmoveFdFlags

FobtainFdFlags

MapFlags

MremapFlags

MunmapFlags

PtraceFlags

RecvFdFlags

RwFlags

SendFdFlags

SigcontrolFlags

Enums

ContextStatus

ContextVerb

FsCall

ProcSchemeVerb

SchemeSocketCall

Constants

ADDRSPACE_OP_MMAP

ADDRSPACE_OP_MPROTECT

ADDRSPACE_OP_MUNMAP

ADDRSPACE_OP_TRANSFER

APPEND

CLOCK_MONOTONIC

CLOCK_REALTIME

CONSUME

Remove the fd from the caller’s file table before sending the message.

EVENT_NONE

EVENT_READ

EVENT_WRITE

FD

Indicates the request is a bulk fd passing request.

FD_CLONE

FD_EXCLUSIVE

Flags for the fd passing request.

FD_UPPER

FIXED

FIXED_REPLACE

FUTEX_REQUEUE

FUTEX_WAIT

FUTEX_WAIT64

FUTEX_WAKE

F_DUPFD

F_GETFD

F_GETFL

F_SETFD

F_SETFL

INHIBIT_DELIVERY

Prevents the kernel from jumping the context to the signal trampoline, but otherwise has absolutely no effect on which signals are blocked etc. Meant to be used for short-lived critical sections inside libc.

KEEP_OLD

Alias’s memory region at old_address to new_address such that both regions share the same frames.

MAP_FIXED

MAP_FIXED_NOREPLACE

MAP_LAZY

For userspace-backed mmaps, return from the mmap call before all pages have been provided by the scheme. This requires the scheme to be trusted, as the current context can block indefinitely, if the scheme does not respond to the page fault handler’s request, as it tries to map the page by requesting it from the scheme.

MAP_PRIVATE

MAP_SHARED

MODE_CHR

MODE_DIR

MODE_FIFO

MODE_FILE

MODE_PERM

MODE_SETGID

MODE_SETUID

MODE_SOCK

MODE_SYMLINK

MODE_TYPE

NEEDS_SYNC

Indicates whether the funmap call must implicitly do an msync, for the changes to become visible later.

NONBLOCK

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECTORY

O_EXCL

O_EXLOCK

O_FSYNC

O_NOFOLLOW

O_NONBLOCK

O_RDONLY

O_RDWR

O_SHLOCK

O_STAT

O_SYMLINK

O_TRUNC

O_WRONLY

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

PTRACE_EVENT_ADDRSPACE_SWITCH

Sent when current-addrspace is changed, allowing the tracer to reopen the memory file.

PTRACE_EVENT_CLONE

Sent when a child is cloned, giving you the opportunity to trace it. If you don’t catch this, the child is started as normal.

PTRACE_EVENT_MASK

PTRACE_FLAG_IGNORE

Special meaning, depending on the event. Usually, when fired before an action, it will skip performing that action.

PTRACE_FLAG_MASK

PTRACE_STOP_BREAKPOINT

Stop on a software breakpoint, such as the int3 instruction for x86_64.

PTRACE_STOP_EXIT

Stop just before exiting for good.

PTRACE_STOP_MASK

PTRACE_STOP_POST_SYSCALL

Stop after a syscall is handled.

PTRACE_STOP_PRE_SYSCALL

Stop before a syscall is handled. Send PTRACE_FLAG_IGNORE to not handle the syscall.

PTRACE_STOP_SIGNAL

Stop before a signal is handled. Send PTRACE_FLAG_IGNORE to not handle signal.

PTRACE_STOP_SINGLESTEP

Stop after exactly one instruction. TODO: This may not handle fexec/signal boundaries. Should it?

READ

RSVD0

RSVD1

RSVD2

RSVD3

RSVD4

RSVD5

RSVD6

RSVD7

SEEK_CUR

SEEK_END

SEEK_SET

SIGCHLD

SIGTSTP

SIGTTIN

SIGTTOU

SKMSG_FOBTAINFD

SKMSG_FRETURNFD

SKMSG_PROVIDE_MMAP

UPPER_FDTBL_TAG

The tag for the fd number in the upper file descriptor table.

WRITE